Facility Readings
-modified ultrasonic field to float
-modified api to get last 3 readings with status (system monitoring(

+++ facility/models.py
@@ -39,7 +39,7 @@ class FacilityReading(BaseModel):
     pir = models.BooleanField()
     temperature = models.IntegerField()
     light = models.IntegerField()
-    ultrasonic = models.IntegerField()
+    ultrasonic = models.FloatField()
     facility = models.ForeignKey(Facility, null=True, on_delete=models.SET_NULL)
 
     objects = FacilityReadingManager()
